# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/invite": {
      "post": {
        "operationId": "UsersController_invite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/tasks": {
      "get": {
        "operationId": "TasksController_findAll",
        "summary": "List all tasks",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "uuid-project",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "todo",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "example": "painting",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create new task",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/{id}": {
      "patch": {
        "operationId": "TasksController_update",
        "summary": "Update task details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/workflows/{id}": {
      "patch": {
        "operationId": "WorkflowsController_updateRule",
        "summary": "Update workflow rule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Workflows"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM Backend API",
    "description": "The core API for CRM Lead Management and Project Coordination.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crm_session",
        "description": "Session cookie (HttpOnly)"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "rid": {
            "type": "string",
            "example": "s-success-rid"
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "type": "object",
            "example": {
              "timestamp": "2026-03-28T07:44:00Z"
            }
          }
        },
        "required": [
          "rid",
          "statusCode",
          "data",
          "message",
          "meta"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Final Painting - Living Room",
            "description": "The title of the task"
          },
          "projectId": {
            "type": "string",
            "example": "uuid-project",
            "description": "Linked Project ID"
          },
          "phaseId": {
            "type": "string",
            "example": "uuid-phase"
          },
          "milestoneId": {
            "type": "string",
            "example": "uuid-milestone"
          },
          "description": {
            "type": "string",
            "example": "Wall painting for the living room including ceiling"
          },
          "status": {
            "type": "string",
            "example": "todo",
            "description": "Initial status of the task"
          },
          "priority": {
            "type": "string",
            "example": "high",
            "description": "Priority level"
          },
          "assignedTo": {
            "type": "string",
            "example": "uuid-user"
          },
          "assigneeId": {
            "type": "string",
            "example": "uuid-user",
            "description": "Alias for assigned user"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-04-01T08:00:00.000Z"
          },
          "estimatedHours": {
            "type": "number",
            "example": 8
          },
          "parentTaskId": {
            "type": "string",
            "example": "uuid-parent-task"
          }
        },
        "required": [
          "title",
          "projectId"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "in_progress"
          },
          "priority": {
            "type": "string",
            "example": "critical"
          },
          "assignedTo": {
            "type": "string",
            "example": "uuid-user"
          },
          "assigneeId": {
            "type": "string",
            "example": "uuid-user"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-04-01T08:00:00.000Z"
          },
          "estimatedHours": {
            "type": "number",
            "example": 8
          },
          "parentTaskId": {
            "type": "string",
            "example": "uuid-parent-task"
          }
        }
      },
...
      "UpdateWorkflowRuleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}